hotfix

+++ actions/scripts/rebuild_app.sh
@@ -9,16 +9,21 @@ counter=0
 if [ -d "$working_dir" ]; then
   cd $working_dir
   ids=$(sudo docker container ls | grep -E "${ptn}" | awk '{print $1}')
-  for i in $ids;
-  do
-    sudo docker container stop $i \
-    && sudo docker container rm $i;
-    counter=`expr $counter + 1`
-  done
+  test=$(echo ${ids:="unknown"})
+  if [ $test = "unknown" ]; then
+    exit 201
+  else
+    for i in $ids;
+    do
+      sudo docker container stop $i \
+      && sudo docker container rm $i;
+      counter=`expr $counter + 1`
+    done
 
-  if [ $counter -eq 2 ]; then
-    sudo docker-compose up -d --build
-    counter=`expr $counter + 1`
+    if [ $counter -eq 2 ]; then
+      sudo docker-compose up -d --build
+      counter=`expr $counter + 1`
+    fi
   fi
 fi
 
@@ -27,10 +32,3 @@ if [ $counter -eq 3 ]; then
 fi
 
 exit 1
-
-
-
-
-
-
-
